[backend] Move Askbot backend to core sub-package
